Setup project folder structure
