--- knexfile.js
@@ -1,17 +1,0 @@
-module.exports = {
-  development: {
-    client: 'sqlite3',
-    connection: { filename: './data/lambda.db3' },
-    useNullAsDefault: true,
-    pool: {
-      afterCreate: (conn, done) => {
-        conn.run('PRAGMA foreign_keys = ON', done);
-      }
-    },
-    migrations: {
-      directory: './data/migrations',
-      tableName: 'dbmigrations',
-    },
-    seeds: { directory: './data/seeds' },
-  },
-};
